third_party/blink/web_tests/http/tests/images/drag-image-to-desktop-expected.txt

CONSOLE MESSAGE: Test image loaded.
CONSOLE MESSAGE: Drop event fired.
Filename being dragged: generatedimage.gif
